move-lit-component-files
commite to move-lit-component-files
