+++ package.json
@@ -16,10 +16,14 @@
   },
   "dependencies": {
     "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-decorators": "^7.23.9",
     "@disciple.tools/web-components": "^0.5.2",
+    "@spectrum-css/popover": "^6.1.0",
     "@spectrum-web-components/action-group": "^0.40.5",
     "@spectrum-web-components/action-menu": "^0.41.0",
     "@spectrum-web-components/button": "^0.40.5",
+    "@spectrum-web-components/button-group": "^0.40.5",
+    "@spectrum-web-components/link": "^0.40.4",
     "@spectrum-web-components/menu": "^0.40.5",
     "@spectrum-web-components/progress-circle": "^0.40.5",
     "element-internals-polyfill": "^1.3.10",
