+++ resources/js/components/app-icon.js
