+++ resources/js/components/menu.js
