+++ resources/js/components/translate.js
