+++ resources/js/plugin.js
@@ -1,9 +1,8 @@
 import "../css/plugin.css";
 import "@disciple.tools/web-components";
-import "./app-grid.js";
-import  "./hidden-app-grid.js";
-
-import "./action-group.js";
-import "./home-footer.js";
-import "./menu.js";
-import "./training-video.js"
+import "./components/app-grid.js";
+import  "./components/hidden-app-grid.js";
+import "./components/action-group.js";
+import "./components/home-footer.js";
+import "./components/menu.js";
+import "./components/training-video.js"
